+++ modules/juce_core/native/juce_SystemStats_mac.mm
@@ -311,7 +311,7 @@ public:
         }
 
         highResTimerFrequency = (timebase.denom * (uint64) 1000000000) / timebase.numer;
-        highResTimerToMillisecRatio = hiResCounterNumerator / (double) hiResCounterDenominator;
+        highResTimerToMillisecRatio = (double) hiResCounterNumerator / (double) hiResCounterDenominator;
     }
 
     uint32 millisecondsSinceStartup() const noexcept
@@ -321,7 +321,7 @@ public:
 
     double getMillisecondCounterHiRes() const noexcept
     {
-        return mach_absolute_time() * highResTimerToMillisecRatio;
+        return (double) mach_absolute_time() * highResTimerToMillisecRatio;
     }
 
     int64 highResTimerFrequency;
