+++ modules/juce_video/native/juce_CameraDevice_ios.h
@@ -836,7 +836,7 @@ private:
                             break;
                         case kCGImagePropertyOrientationRight:
                             CGContextRotateCTM (context, 90 * MathConstants<CGFloat>::pi / 180);
-                            CGContextScaleCTM (context, targetSize.height / origHeight, -targetSize.width / origWidth);
+                            CGContextScaleCTM (context, targetSize.height / (CGFloat) origHeight, -targetSize.width / (CGFloat) origWidth);
                             break;
                         case kCGImagePropertyOrientationDown:
                             CGContextTranslateCTM (context, targetSize.width, 0.0);
@@ -844,7 +844,7 @@ private:
                             break;
                         case kCGImagePropertyOrientationLeft:
                             CGContextRotateCTM (context, -90 * MathConstants<CGFloat>::pi / 180);
-                            CGContextScaleCTM (context, targetSize.height / origHeight, -targetSize.width / origWidth);
+                            CGContextScaleCTM (context, targetSize.height / (CGFloat) origHeight, -targetSize.width / (CGFloat) origWidth);
                             CGContextTranslateCTM (context, -targetSize.width, -targetSize.height);
                             break;
                         case kCGImagePropertyOrientationUpMirrored:
